requirements.txt
autopep8==1.6.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cycler==0.11.0
filelock==3.1.0
fonttools==4.28.4
huggingface-hub==0.0.17
idna==3.2
joblib==1.0.1
kiwisolver==1.3.2
matplotlib==3.5.1
mysql-connector-python==8.0.26
nltk==3.6.3
numpy==1.21.2
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.0
pycodestyle==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
regex==2021.9.24
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.16.0
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
transformers==4.11.0
typing-extensions==3.10.0.2
urllib3==1.26.7
